requirements.txt
asgiref==3.5.0
certifi==2021.10.8
charset-normalizer==2.0.12
Django==4.0.4
django-filter==21.1
djangorestframework==3.13.1
docopt==0.6.2
idna==3.3
importlib-metadata==4.11.3
Markdown==3.3.6
pipreqs==0.4.11
pytz==2022.1
requests==2.27.1
sqlparse==0.4.2
urllib3==1.26.9
yarg==0.1.9
zipp==3.8.0
